+++ core/src/Streamly/Internal/Data/ChunkParserK.hs
@@ -1,5 +1,5 @@
 -- |
--- Module      : Streamly.Internal.Data.Parser.ParserK.Type
+-- Module      : Streamly.Internal.Data.ChunkParserK
 -- Copyright   : (c) 2020 Composewell Technologies
 -- License     : BSD-3-Clause
 -- Maintainer  : streamly@composewell.com
@@ -21,12 +21,12 @@
 -- <http://hackage.haskell.org/package/parser-combinators parser-combinators>
 -- package can also be used with the 'ParserK' type.
 
-module Streamly.Internal.Data.Parser.ParserK.Type
+module Streamly.Internal.Data.ChunkParserK
     (
       Step (..)
     , Input (..)
     , ParseResult (..)
-    , ParserK (..)
+    , ChunkParserK (..)
     , fromParser
     -- , toParser
     , fromPure
@@ -118,7 +118,7 @@ instance Functor ParseResult where
 -- input without a result or return a result with no further input to be
 -- consumed.
 --
-newtype ParserK a m b = MkParser
+newtype ChunkParserK a m b = MkParser
     { runParser :: forall r.
            -- Using "Input" in runParser is not necessary but it avoids making
            -- one more function call to get the input. This could be helpful
@@ -156,7 +156,7 @@ newtype ParserK a m b = MkParser
 -- XXX rewrite this using ParserD, expose rmapM from ParserD.
 -- | Maps a function over the output of the parser.
 --
-instance Functor m => Functor (ParserK a m) where
+instance Functor m => Functor (ChunkParserK a m) where
     {-# INLINE fmap #-}
     fmap f parser = MkParser $ \k n st arr ->
         let k1 res = k (fmap f res)
@@ -173,7 +173,7 @@ instance Functor m => Functor (ParserK a m) where
 -- /Pre-release/
 --
 {-# INLINE fromPure #-}
-fromPure :: b -> ParserK a m b
+fromPure :: b -> ChunkParserK a m b
 fromPure b = MkParser $ \k n st arr -> k (Success n b) st arr
 
 -- | See 'Streamly.Internal.Data.Parser.fromEffect'.
@@ -181,7 +181,7 @@ fromPure b = MkParser $ \k n st arr -> k (Success n b) st arr
 -- /Pre-release/
 --
 {-# INLINE fromEffect #-}
-fromEffect :: Monad m => m b -> ParserK a m b
+fromEffect :: Monad m => m b -> ChunkParserK a m b
 fromEffect eff =
     MkParser $ \k n st arr -> eff >>= \b -> k (Success n b) st arr
 
@@ -189,7 +189,7 @@ fromEffect eff =
 -- this operation does not fuse, use 'Streamly.Internal.Data.Parser.splitWith'
 -- when fusion is important.
 --
-instance Monad m => Applicative (ParserK a m) where
+instance Monad m => Applicative (ChunkParserK a m) where
     {-# INLINE pure #-}
     pure = fromPure
 
@@ -226,7 +226,7 @@ instance Monad m => Applicative (ParserK a m) where
 -- /Pre-release/
 --
 {-# INLINE die #-}
-die :: String -> ParserK a m b
+die :: String -> ChunkParserK a m b
 die err = MkParser (\k n st arr -> k (Failure n err) st arr)
 
 -- | Monad composition can be used for lookbehind parsers, we can make the
@@ -271,7 +271,7 @@ die err = MkParser (\k n st arr -> k (Failure n err) st arr)
 -- does not fuse, use 'Streamly.Internal.Data.Parser.concatMap' when you need
 -- fusion.
 --
-instance Monad m => Monad (ParserK a m) where
+instance Monad m => Monad (ChunkParserK a m) where
     {-# INLINE return #-}
     return = pure
 
@@ -290,11 +290,11 @@ instance Monad m => Monad (ParserK a m) where
     {-# INLINE fail #-}
     fail = die
 #endif
-instance Monad m => Fail.MonadFail (ParserK a m) where
+instance Monad m => Fail.MonadFail (ChunkParserK a m) where
     {-# INLINE fail #-}
     fail = die
 
-instance MonadIO m => MonadIO (ParserK a m) where
+instance MonadIO m => MonadIO (ChunkParserK a m) where
     {-# INLINE liftIO #-}
     liftIO = fromEffect . liftIO
 
@@ -315,7 +315,7 @@ instance MonadIO m => MonadIO (ParserK a m) where
 -- does not fuse, use 'Streamly.Internal.Data.Parser.alt' when you need
 -- fusion.
 --
-instance Monad m => Alternative (ParserK a m) where
+instance Monad m => Alternative (ChunkParserK a m) where
     {-# INLINE empty #-}
     empty = die "empty"
 
@@ -348,7 +348,7 @@ instance Monad m => Alternative (ParserK a m) where
 -- | 'mzero' is same as 'empty', it aborts the parser. 'mplus' is same as
 -- '<|>', it selects the first succeeding parser.
 --
-instance Monad m => MonadPlus (ParserK a m) where
+instance Monad m => MonadPlus (ChunkParserK a m) where
     {-# INLINE mzero #-}
     mzero = die "mzero"
 
@@ -494,7 +494,7 @@ parseDToK pstep initial extract cont !offset0 !usedCount !input = do
 -- /Pre-release/
 --
 {-# INLINE_LATE fromParser #-}
-fromParser :: (Monad m, Unbox a) => ParserD.Parser a m b -> ParserK a m b
+fromParser :: (Monad m, Unbox a) => ParserD.Parser a m b -> ChunkParserK a m b
 fromParser (ParserD.Parser step initial extract) =
     MkParser $ parseDToK step initial extract
 
@@ -515,7 +515,7 @@ parserDone _ _ _ = error "Bug: toParser: called with input"
 -- /Pre-release/
 --
 {-# INLINE_LATE toParser #-}
-toParser :: Monad m => ParserK a m b -> ParserD.Parser a m b
+toParser :: Monad m => ChunkParserK a m b -> ParserD.Parser a m b
 toParser parser = ParserD.Parser step initial extract
 
     where